Fold trello_models.dart into members.dart
