+++ lib/src/sdk/members/members.dart
@@ -1,8 +1,5 @@
 import 'package:trello_sdk/src/sdk/types.dart';
 
-import '../trello_models.dart'
-    show BoardBackgrounds, BoardsInvited, BoardsInvitedFields;
-
 export 'member_board_filter.dart';
 export 'member_client.dart';
 export 'member_fields.dart';
@@ -24,3 +21,45 @@ typedef MemberNotifications = String;
 typedef MemberOrganizationFields = OrganizationFields;
 typedef MemberOrganizationsInvited = OrganizationsInvited;
 typedef MemberOrganizationsInvitedFields = OrganizationsInvitedFields;
+
+/// 'default' is a reserved word, so is here in uppercase
+/// use the names in lowercase to get correct form
+enum BoardBackgrounds {
+  all,
+  custom,
+  defaulT,
+  none,
+  premium;
+}
+
+enum BoardsInvited {
+  all,
+  closed,
+  members,
+  open,
+  organization,
+  pinned,
+  public,
+  starred,
+  unpinned
+}
+
+enum BoardsInvitedFields {
+  all,
+  id,
+  name,
+  desc,
+  descData,
+  closed,
+  idMemberCreator,
+  idOrganization,
+  pinned,
+  url,
+  shortUrl,
+  prefs,
+  labelNames,
+  starred,
+  limits,
+  memberships,
+  enterpriseOwned
+}
